migrated playlist-routes to ts
